types: Add From/Into conversions for Position and GeneratedSpan
These allow simpler construction leveraging common
idioms of `.into()` and `::from()`.

We are unable to implement any for Span because it is
based on an external type so the rust orphan rules
prevent the impl

Adding an impl on a type constraint for Into<ParsedSpan>
conflicts with adding a second impl with a different type
constraint for Into<GeneratedSpan>.  Those are added with
concrete types.

For `Position`:
  * From `Into<ParsedSpan>`
  * From `GeneratedSpan`
  * From `Option<String>`
    (sugar for Option<String> -> GeneratedSpan -> Position )

For `GeneratedSpan`
  * From Option<String>